updateeeeeeeeeeeee

+++ src/main/java/com/eshop/jpaRepository/CartRepository.java
@@ -7,10 +7,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CartRepository extends JpaRepository<Cart, Long> {
-    public List<Cart> findByUserId(int userId);
-    public List<Cart> findByUserIdAndProductId(int userId, int slug);
-    public Boolean existsShoppingCartByUserIdAndProductId(int userId, int productId);
-    public Cart findByProductId(int productId);
-    public Cart deleteByProductId(int productId);
+public interface CartRepository extends JpaRepository<Cart, Integer> {
+    public List<Cart> findByUserId(Integer userId);
+    public List<Cart> findByUserIdAndProductId(Integer userId, Integer slug);
+    public Boolean existsShoppingCartByUserIdAndProductId(Integer userId, Integer productId);
+    public Cart findByProductId(Integer productId);
+    public Cart deleteByProductId(Integer productId);
 }
